Add selective structural indexing

+++ src/services/indexer.js
@@ -18,7 +18,7 @@ class State {
   /** @type {STATE} */
   state;
   count = 0;
-  constructor() {}
+  constructor() { }
   get isIndexed() {
     return this.count > 0;
   }
@@ -96,21 +96,25 @@ export async function getState(type) {
   return states[type];
 }
 
-export async function deleteIndex(type) {
+export async function deleteIndex(type, crateId) {
   if (!states[type] || states[type].isBusy) return;
   states[type].startDeleting();
-  await indexer[type]?.delete();
+  await indexer[type]?.delete(crateId);
   states[type].stopDeleting();
 }
 
-export async function createIndex(type, force = false) {
+export async function createIndex(type, crateId, force = false) {
   if (!states[type] || states[type].isBusy) return;
-  if (force) {
+  if (force && !crateId) {
     await deleteIndex(type);
   }
   states[type].startIndexing();
-  await indexRepository({ repository: repo, skipByMatch, types: [type] });
+  console.log('Indexing started');
+  const counts = await indexRepository({ repository: repo, skipByMatch, types: [type], crateId });
+  console.log('Indexing stopped');
+  console.log(counts);
   states[type].stopIndexing();
+  return true;
 }
 
 /**
@@ -119,28 +123,35 @@ export async function createIndex(type, force = false) {
  * @param {string[]} [p.types]
  * @param {import('@ocfl/ocfl').OcflStorage} p.repository
  * @param {string[]} [p.skipByMatch]
+ * @param {string} [p.crateId]
  */
-export async function indexRepository({ repository, skipByMatch = [], types = [] }) {
+export async function indexRepository({ repository, skipByMatch = [], types = [], crateId }) {
   const skipRegExp = skipByMatch.length ? new RegExp(skipByMatch.join("|"), "i") : null;
   if (!types.length) return;
-  let counts = Object.fromEntries(types.map(t => [t, 0]));
+  const counts = Object.fromEntries(types.map(t => [t, 0]));
   for await (const ocflObject of repository) {
-    await ocflObject.load();
-    if (skipRegExp && skipRegExp.test(ocflObject.id)) {
+    if (skipRegExp?.test(ocflObject.id)) {
       logger.warn(`${ocflObject.id} : Skipping object by match in config: ${skipRegExp}`);
     } else {
       logger.debug(`Loading record: ${ocflObject.root}`);
       try {
-        const jsonContent = await ocflObject.getFile({ logicalPath: 'ro-crate-metadata.json' }).asString();
+        //await ocflObject.load();
+        const jsonContent = await ocflObject.getFile({ logicalPath: 'ro-crate-metadata.json' }).text();
         const rawCrate = JSON.parse(jsonContent);
         const crate = new ROCrate(rawCrate, { array: true, link: true });
-        for (const t of types) {
-          try {
-            await indexer[t].index({ ocflObject, crate });
-            counts[t]++;
-          } catch (error) {
+        // if crateId is specified, index just the object and the subcollections and child objects
+        // by checking just the structure implied in the crate id as a hack
+        if (!crateId || crate.rootId.startsWith(crateId)) {
+          for (const t of types) {
+            try {
+              if (crateId) await indexer[t].delete(crateId);
+              await indexer[t].index({ ocflObject, crate });
+              counts[t]++;
+            } catch (error) {
+              logger.error(error.message);
+            }
           }
-        }
+        }        
       } catch (e) {
         logger.error(e.message);
       }
@@ -148,5 +159,3 @@ export async function indexRepository({ repository, skipByMatch = [], types = []
   }
   return counts;
 }
-
-
